Imports works with basic schema
#2
As a part of this commit I refactored code for Selection panel so it can
be used from all views in the project, also refactored code on the views
to use helper classes.
The import will work in basic (and probably most useful) situation.
Still need to check some edge situations